requirements.txt
gspread>=6.0.0
google-auth>=2.0.0
pyTelegramBotAPI>=4.14.0
python-dotenv>=1.0.0
APScheduler>=3.10.4
